Move register_translation from feature to registration
