+++ skimage/feature/tests/test_masked_register_translation.py
@@ -4,7 +4,7 @@ from scipy.ndimage import fourier_shift
 from skimage._shared import testing
 from skimage._shared.testing import assert_equal
 from skimage.data import camera
-from skimage.feature.register_translation import register_translation
+from skimage.registration.register_translation import register_translation
 from skimage.feature.masked_register_translation import (
     masked_register_translation, cross_correlate_masked)
 from skimage.io import imread
