+++ skimage/registration/__init__.py
@@ -1,1 +1,8 @@
 from ._optical_flow import optical_flow_tvl1
+from .register_translation import register_translation
+
+
+__all__ = [
+    'optical_flow_tvl1',
+    'register_translation'
+    ]
