+++ skimage/registration/register_translation.py
